Split primitives into separate files and fix build
