+++ modules/world/tests/BUILD
@@ -132,7 +132,8 @@ cc_test(
         "//modules/world/opendrive:opendrive",
         "//modules/models/behavior/constant_velocity:constant_velocity",
         "//modules/models/execution/interpolation:interpolation",
-        "//modules/world/evaluation:evaluation",        
+        "//modules/world/evaluation:evaluation",
+        "//modules/models/behavior/motion_primitives:motion_primitives",
         ":make_test_world",
         ":dummy_road_corridor",
         "@gtest//:gtest_main",
